はい、大丈夫です

+++ src/main/java/com/rpg/fried_shrimp/controller/BattleController.java
@@ -28,7 +28,7 @@ public interface BattleController {
     
     // 再挑戦用
     //@GetMapping("/battle/retry/{enemyId}")
-    public String retryBattle(@PathVariable Long enemyId, Model model);
+    public String retryBattle(@PathVariable int enemyId, Model model);
     
     // 次のバトルへ
     //@PostMapping("/battle/next")
